% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqc.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{PcjProcessCapabilityModel1}
\alias{PcjSequentialProcedure}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pcj}{\code{\link[pcj]{PcjProcessCapabilityModel1}}, \code{\link[pcj]{PcjSequentialProcedure}}}
}}

